WIP refactoring

+++ firmware/firmware/MIDIConfig.h
@@ -12,10 +12,10 @@ public:
     MidiConfig() : _channel(DEFAULT_MIDI_CHANNEL), _octave(DEFAULT_OCTAVE) {}
     virtual ~MidiConfig() = default;
 
-    int getChannel() { return _channel; };
-    void setChannel(int c) { _channel = c; };
-    int getOctave() { return _octave; };
-    void setOctave(int o) { _octave = o; };
+    uint8_t getChannel() { return _channel; };
+    void setChannel(uint8_t c) { _channel = c; };
+    uint8_t getOctave() { return _octave; };
+    void setOctave(uint8_t o) { _octave = o; };
 
 private:
     uint8_t _channel;
